requirements.txt
requests
python-telegram-bot
googletrans==4.0.0-rc1
# otras dependencias...